requirements.txt
Flask==2.2.3
tensorflow==2.12.0
tensorflow-hub
numpy
Pillow
matplotlib
requests
gunicorn
gevent